# server/swagger.yaml
openapi: 3.0.0
info:
  title: Agentic Schedule API
  version: 1.0.0
  description: API documentation for Contracts, Staff, Skills, DayOffRequests, and ShiftOffRequests
paths:
...
  /shiftoffrequests:
    post:
      summary: Create one or multiple shift off requests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ShiftOffRequest'
                - type: array
                  items:
                    $ref: '#/components/schemas/ShiftOffRequest'
      responses:
        '201':
          description: Shift off requests created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 201
                  success_message:
                    type: string
                    example: Shift off requests created successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      requestIds:
                        type: array
                        items:
                          type: integer
                        example: [1, 2, 3]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    get:
      summary: Get all shift off requests
      responses:
        '200':
          description: List of shift off requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Shift off requests fetched successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      shiftOffRequests:
                        type: array
                        items:
                          $ref: '#/components/schemas/ShiftOffRequest'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    delete:
      summary: Delete shift off requests by list of RequestID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requestIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
      responses:
        '200':
          description: Shift off requests deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Shift off requests deleted successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
  /requesttype:
    post:
      summary: Create one or multiple request types
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/RequestType'
                - type: array
                  items:
                    $ref: '#/components/schemas/RequestType'
            examples:
              SingleRequestType:
                summary: Single request type object
                value:
                  RequestTypeName: "Vacation"
                  RequestTypeDesc: "Vacation leave"
              MultipleRequestTypes:
                summary: List of request type objects
                value:
                  - RequestTypeName: "Sick"
                    RequestTypeDesc: "Sick leave"
                  - RequestTypeName: "Personal"
                    RequestTypeDesc: "Personal leave"
      responses:
        '201':
          description: Request types created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 201
                  success_message:
                    type: string
                    example: Request types created successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      requestTypeIds:
                        type: array
                        items:
                          type: integer
                        example: [1, 2]
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    get:
      summary: Get all request types
      responses:
        '200':
          description: List of request types
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Request types fetched successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      requestTypes:
                        type: array
                        items:
                          $ref: '#/components/schemas/RequestType'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
    delete:
      summary: Delete request types by list of RequestTypeID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requestTypeIds:
                  type: array
                  items:
                    type: integer
                  example: [1, 2]
      responses:
        '200':
          description: Request types deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: Request types deleted successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      deletedCount:
                        type: integer
                        example: 2
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
  /requesttype/{id}:
    get:
      summary: Get request type by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Request type
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  statusCode:
                    type: integer
                    example: 200
                  success_message:
                    type: string
                    example: RequestType fetched successfully
                  error_message:
                    type: string
                    example: ""
                  payload:
                    type: object
                    properties:
                      requestType:
                        $ref: '#/components/schemas/RequestType'
        '404':
          description: Request type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
...
components:
  schemas:
...
    ShiftOffRequest:
      type: object
      properties:
        employeeId:
          type: integer
        requestDate:
          type: string
          format: date
        requestTypeId:
          type: integer
        shift:
          type: string
    RequestType:
      type: object
      properties:
        RequestTypeID:
          type: integer
        RequestTypeName:
          type: string
        RequestTypeDesc:
          type: string
...
    StandardErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        statusCode:
          type: integer
          example: 400
        success_message:
          type: string
          example: ""
        error_message:
          type: string
          example: "Invalid data"
        payload:
          type: object
          example: {}
...